+++ Gems/Atom/Tools/AtomToolsFramework/Code/Include/AtomToolsFramework/Graph/GraphViewSettings.h
@@ -9,9 +9,9 @@
 #pragma once
 
 #if !defined(Q_MOC_RUN)
+#include <AtomToolsFramework/Graph/GraphViewConstructPresets.h>
 #include <GraphCanvas/Editor/AssetEditorBus.h>
 #include <GraphCanvas/Widgets/NodePalette/TreeItems/NodePaletteTreeItem.h>
-#include <AtomToolsFramework/GraphView/GraphViewConstructPresets.h>
 #endif
 
 namespace AtomToolsFramework
