+++ Gems/Atom/Tools/AtomToolsFramework/Code/Source/Graph/DynamicNode/DynamicNodeConfig.cpp
@@ -6,8 +6,8 @@
  *
  */
 
-#include <AtomToolsFramework/DynamicNode/DynamicNodeConfig.h>
-#include <AtomToolsFramework/DynamicNode/DynamicNodeUtil.h>
+#include <AtomToolsFramework/Graph/DynamicNode/DynamicNodeConfig.h>
+#include <AtomToolsFramework/Graph/DynamicNode/DynamicNodeUtil.h>
 #include <AtomToolsFramework/Util/Util.h>
 #include <AzCore/RTTI/BehaviorContext.h>
 #include <AzCore/Serialization/EditContext.h>
