+++ Gems/Atom/Tools/MaterialCanvas/Code/Source/MaterialCanvasTestData.cpp
@@ -7,7 +7,7 @@
  */
 
 #include <Atom/RPI.Edit/Common/AssetUtils.h>
-#include <AtomToolsFramework/DynamicNode/DynamicNodeConfig.h>
+#include <AtomToolsFramework/Graph/DynamicNode/DynamicNodeConfig.h>
 #include <AzCore/IO/FileIO.h>
 #include <AzCore/IO/Path/Path.h>
 #include <AzCore/std/algorithm.h>
